Move the class to its native package. Make it package-private.
